dom/tests/mochitest/bugs/test_bug740811.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=740811
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 740811</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=740811">Mozilla Bug 740811</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 740811 **/
is(Object.getPrototypeOf(SVGElement.prototype), Element.prototype,
   "Bogus proto for SVGElement");
</script>
</pre>
</body>
</html>
